+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to
 - 🐛(migrations) use settings in migrations #1058
 - 💄(frontend) truncate pinned participant name with ellipsis on overflow #1056
 - ♿(frontend) prevent focus ring clipping on invite dialog #1078
+- ♿(frontend) dynamic tab title when connected to meeting #1060
 
 ## [1.9.0] - 2026-03-02